# div_mod_knuth added