Added div_mod_knuth division for two large >128bit numbers. Resolves #16.