@mcseemk 2022-07-07
Sure,
bench before:

U256::add               time:   [1.6429 ns 1.6462 ns 1.6499 ns]                       
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) low mild
  7 (7.00%) high mild
  1 (1.00%) high severe

U256::div (lo/lo)       time:   [8.1225 ns 8.1331 ns 8.1452 ns]                               
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) high mild
  10 (10.00%) high severe

U256::div (hi/lo)       time:   [19.067 ns 19.083 ns 19.100 ns]                               
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild

U256::div (hi/hi)       time:   [171.46 ns 172.14 ns 172.87 ns]                              

U256::mul               time:   [2.6050 ns 2.6064 ns 2.6078 ns]                       
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe

U256::sub               time:   [1.6185 ns 1.6207 ns 1.6232 ns]                       
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

U256::shl               time:   [1.9257 ns 1.9313 ns 1.9371 ns]                       
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

U256::shr               time:   [1.7386 ns 1.7402 ns 1.7424 ns]                       
Found 17 outliers among 100 measurements (17.00%)
  4 (4.00%) high mild
  13 (13.00%) high severe

U256::ctlz              time:   [798.04 ps 798.59 ps 799.28 ps]                        
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  8 (8.00%) high severe

U256::cttz              time:   [881.57 ps 882.25 ps 883.15 ps]                        
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

U256::rotate_left       time:   [2.9088 ns 2.9195 ns 2.9321 ns]                               
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

U256::rotate_right      time:   [2.7154 ns 2.7197 ns 2.7249 ns]                                
Found 14 outliers among 100 measurements (14.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  10 (10.00%) high severe


and after:


U256::add               time:   [1.6201 ns 1.6232 ns 1.6267 ns]                       
                        change: [-1.5315% -1.1987% -0.8861%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

U256::div (lo/lo)       time:   [8.1195 ns 8.1333 ns 8.1489 ns]                               
                        change: [-0.2028% -0.0507% +0.1022%] (p = 0.52 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe

U256::div (hi/lo)       time:   [19.281 ns 19.295 ns 19.309 ns]                               
                        change: [+1.0802% +1.2256% +1.3774%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

U256::div (hi/hi)       time:   [24.055 ns 24.095 ns 24.141 ns]                               
                        change: [-86.074% -86.006% -85.935%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  2 (2.00%) high severe

U256::mul               time:   [2.6083 ns 2.6114 ns 2.6154 ns]                       
                        change: [+0.3169% +0.7328% +1.2796%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

U256::sub               time:   [1.6206 ns 1.6233 ns 1.6262 ns]                       
                        change: [-0.0591% +0.1864% +0.4386%] (p = 0.15 > 0.05)
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::shl               time:   [1.9175 ns 1.9225 ns 1.9283 ns]                       
                        change: [-0.6044% -0.1814% +0.2108%] (p = 0.39 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

U256::shr               time:   [1.7385 ns 1.7405 ns 1.7439 ns]                       
                        change: [-0.4575% -0.1798% +0.0966%] (p = 0.21 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe

U256::ctlz              time:   [799.83 ps 801.46 ps 803.33 ps]                        
                        change: [+0.8143% +1.2245% +1.6966%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

U256::cttz              time:   [884.29 ps 885.95 ps 887.80 ps]                        
                        change: [+0.1541% +0.3471% +0.5903%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

U256::rotate_left       time:   [2.9030 ns 2.9095 ns 2.9165 ns]                               
                        change: [-0.4245% -0.1878% +0.0451%] (p = 0.13 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  9 (9.00%) high severe

U256::rotate_right      time:   [2.7006 ns 2.7032 ns 2.7068 ns]                                
                        change: [-0.6416% -0.4652% -0.3043%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe