@mcseemk 2022-07-08
Previous post was on my desktop PC Intel i12900K. Below is the Graviton 3 (c7g amazon aws instance) bench:
Before:
U256::add               time:   [5.4664 ns 5.4688 ns 5.4718 ns]                       
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  1 (1.00%) high severe

U256::div (lo/lo)       time:   [20.475 ns 20.477 ns 20.479 ns]                               
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) low severe
  10 (10.00%) low mild
  7 (7.00%) high mild
  1 (1.00%) high severe

U256::div (hi/lo)       time:   [94.612 ns 94.762 ns 94.916 ns]                              

U256::div (hi/hi)       time:   [336.59 ns 336.61 ns 336.63 ns]                              
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low severe
  5 (5.00%) high mild
  1 (1.00%) high severe

U256::mul               time:   [5.5886 ns 5.5908 ns 5.5931 ns]                       
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild

U256::sub               time:   [5.4582 ns 5.4601 ns 5.4619 ns]                       
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  4 (4.00%) high mild

U256::shl               time:   [4.2159 ns 4.2164 ns 4.2169 ns]                       
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  4 (4.00%) high severe

U256::shr               time:   [4.1342 ns 4.1354 ns 4.1369 ns]                       
Found 13 outliers among 100 measurements (13.00%)
  7 (7.00%) high mild
  6 (6.00%) high severe

U256::ctlz              time:   [2.4419 ns 2.4420 ns 2.4420 ns]                        
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

U256::cttz              time:   [2.4857 ns 2.4858 ns 2.4858 ns]                        
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::rotate_left       time:   [5.7390 ns 5.7577 ns 5.7763 ns]                               
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

U256::rotate_right      time:   [5.5316 ns 5.5386 ns 5.5461 ns]                                
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe



After:


U256::add               time:   [5.4514 ns 5.4521 ns 5.4529 ns]                       
                        change: [-0.3786% -0.3263% -0.2762%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) low severe
  4 (4.00%) low mild
  1 (1.00%) high severe

U256::div (lo/lo)       time:   [20.474 ns 20.480 ns 20.485 ns]                               
                        change: [-0.0804% -0.0553% -0.0291%] (p = 0.00 < 0.05)
                        Change within noise threshold.

U256::div (hi/lo)       time:   [94.189 ns 94.202 ns 94.215 ns]                              
                        change: [-0.7320% -0.6301% -0.5264%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild

U256::div (hi/hi)       time:   [68.181 ns 68.192 ns 68.204 ns]                              
                        change: [-79.746% -79.742% -79.738%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::mul               time:   [5.5856 ns 5.5879 ns 5.5902 ns]                       
                        change: [-0.1402% -0.0720% -0.0009%] (p = 0.05 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) low mild
  1 (1.00%) high mild

U256::sub               time:   [5.4708 ns 5.4722 ns 5.4736 ns]                       
                        change: [+0.2187% +0.2696% +0.3182%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild

U256::shl               time:   [4.2157 ns 4.2162 ns 4.2167 ns]                       
                        change: [-0.0453% -0.0197% +0.0044%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  3 (3.00%) high mild

U256::shr               time:   [4.1333 ns 4.1340 ns 4.1348 ns]                       
                        change: [-0.0832% -0.0447% -0.0078%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  3 (3.00%) high severe

U256::ctlz              time:   [2.4419 ns 2.4420 ns 2.4421 ns]                        
                        change: [-0.0089% -0.0009% +0.0079%] (p = 0.84 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  5 (5.00%) high mild
  1 (1.00%) high severe

U256::cttz              time:   [2.4824 ns 2.4834 ns 2.4842 ns]                        
                        change: [-0.2846% -0.2477% -0.2093%] (p = 0.00 < 0.05)
                        Change within noise threshold.

U256::rotate_left       time:   [5.7638 ns 5.7770 ns 5.7908 ns]                               
                        change: [+0.3731% +0.9748% +1.6579%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  4 (4.00%) high severe

U256::rotate_right      time:   [5.5470 ns 5.5548 ns 5.5626 ns]                                
                        change: [-0.1547% +0.0932% +0.3309%] (p = 0.46 > 0.05)
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild