@mcseemk 2022-07-08
Graviton 2 (c6g* amazon aws):
Before:

U256::add               time:   [8.3676 ns 8.3697 ns 8.3718 ns]                       
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

U256::div (lo/lo)       time:   [31.313 ns 31.330 ns 31.345 ns]                               
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::div (hi/lo)       time:   [142.58 ns 142.62 ns 142.66 ns]                              
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

U256::div (hi/hi)       time:   [432.31 ns 432.42 ns 432.55 ns]                              
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

U256::mul               time:   [21.649 ns 21.654 ns 21.661 ns]                       
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

U256::sub               time:   [8.3902 ns 8.3923 ns 8.3945 ns]                       
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

U256::shl               time:   [5.9219 ns 5.9233 ns 5.9248 ns]                       
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

U256::shr               time:   [5.8867 ns 5.8869 ns 5.8871 ns]                       
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high severe

U256::ctlz              time:   [3.2917 ns 3.2931 ns 3.2947 ns]                        
Found 12 outliers among 100 measurements (12.00%)
  10 (10.00%) high mild
  2 (2.00%) high severe

U256::cttz              time:   [3.2852 ns 3.2862 ns 3.2872 ns]                        
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low severe
  1 (1.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

U256::rotate_left       time:   [8.6338 ns 8.6409 ns 8.6480 ns]                               
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

U256::rotate_right      time:   [8.5947 ns 8.5984 ns 8.6023 ns]                                
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild


After:


U256::add               time:   [8.3776 ns 8.3796 ns 8.3818 ns]                       
                        change: [+0.0877% +0.1233% +0.1592%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  1 (1.00%) high severe

U256::div (lo/lo)       time:   [31.350 ns 31.386 ns 31.423 ns]                               
                        change: [+0.2172% +0.3445% +0.4766%] (p = 0.00 < 0.05)
                        Change within noise threshold.

U256::div (hi/lo)       time:   [143.55 ns 143.63 ns 143.72 ns]                              
                        change: [+0.5793% +0.6375% +0.7054%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) low mild
  6 (6.00%) high mild

U256::div (hi/hi)       time:   [132.81 ns 132.88 ns 132.96 ns]                              
                        change: [-69.311% -69.290% -69.270%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

U256::mul               time:   [21.653 ns 21.658 ns 21.664 ns]                       
                        change: [-0.0171% +0.0175% +0.0521%] (p = 0.32 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

U256::sub               time:   [8.3743 ns 8.3750 ns 8.3757 ns]                       
                        change: [-0.1994% -0.1761% -0.1541%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

U256::shl               time:   [5.9165 ns 5.9166 ns 5.9168 ns]                       
                        change: [-0.1599% -0.1259% -0.0897%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  1 (1.00%) high mild
  4 (4.00%) high severe

U256::shr               time:   [5.8862 ns 5.8863 ns 5.8865 ns]                       
                        change: [-0.0270% -0.0121% +0.0038%] (p = 0.13 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low severe
  2 (2.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::ctlz              time:   [3.2861 ns 3.2862 ns 3.2863 ns]                        
                        change: [-0.2516% -0.2076% -0.1648%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe

U256::cttz              time:   [3.2849 ns 3.2857 ns 3.2865 ns]                        
                        change: [-0.0756% -0.0139% +0.0411%] (p = 0.65 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low severe
  2 (2.00%) high mild

U256::rotate_left       time:   [8.6495 ns 8.6562 ns 8.6631 ns]                               
                        change: [+0.0940% +0.2308% +0.3666%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe