@mcseemk 2022-07-08
Intel(R) Xeon(R) CPU E3-1270 v6 @ 3.80GHz
Before:
U256::add               time:   [3.7044 ns 3.7158 ns 3.7301 ns]                       
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe

U256::div (lo/lo)       time:   [34.656 ns 34.746 ns 34.849 ns]                               
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe

U256::div (hi/lo)       time:   [101.12 ns 101.34 ns 101.57 ns]                              
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

U256::div (hi/hi)       time:   [260.59 ns 261.64 ns 263.05 ns]                              
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

U256::mul               time:   [7.2249 ns 7.2486 ns 7.2787 ns]                       
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

U256::sub               time:   [3.6266 ns 3.6375 ns 3.6507 ns]                       
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  5 (5.00%) high severe

U256::shl               time:   [4.6821 ns 4.7187 ns 4.7582 ns]                       
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

U256::shr               time:   [5.4234 ns 5.6343 ns 5.8779 ns]                       
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

U256::ctlz              time:   [1.7979 ns 1.8071 ns 1.8172 ns]                        
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

U256::cttz              time:   [1.6234 ns 1.6275 ns 1.6317 ns]                        
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

U256::rotate_left       time:   [7.6074 ns 7.6350 ns 7.6708 ns]                               
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe

U256::rotate_right      time:   [6.9703 ns 6.9921 ns 7.0137 ns]                                
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild


After:


U256::add               time:   [3.8436 ns 4.0444 ns 4.2842 ns]                       
                        change: [+0.9428% +3.2698% +6.2352%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low mild
  13 (13.00%) high severe

U256::div (lo/lo)       time:   [34.616 ns 34.722 ns 34.836 ns]                               
                        change: [-1.8507% -1.2031% -0.6344%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

U256::div (hi/lo)       time:   [102.35 ns 102.50 ns 102.67 ns]                              
                        change: [+0.9068% +1.1183% +1.3322%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

U256::div (hi/hi)       time:   [76.856 ns 80.566 ns 84.570 ns]                              
                        change: [-71.760% -71.055% -70.233%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  13 (13.00%) high severe

U256::mul               time:   [7.2813 ns 7.3206 ns 7.3653 ns]                       
                        change: [-0.3070% +0.5551% +1.2446%] (p = 0.16 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

U256::sub               time:   [3.6143 ns 3.6161 ns 3.6180 ns]                       
                        change: [-0.8392% -0.5666% -0.3273%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) low severe
  3 (3.00%) high mild

U256::shl               time:   [4.9900 ns 5.2624 ns 5.5863 ns]                       
                        change: [+4.5167% +8.7856% +13.977%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) high mild
  14 (14.00%) high severe

U256::shr               time:   [5.2371 ns 5.3298 ns 5.4537 ns]                       
                        change: [-7.4553% -4.3827% -1.3965%] (p = 0.01 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

U256::ctlz              time:   [1.9459 ns 2.0639 ns 2.2043 ns]                        
                        change: [+6.6873% +11.117% +15.950%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) high mild
  13 (13.00%) high severe

U256::cttz              time:   [1.6350 ns 1.6419 ns 1.6494 ns]                        
                        change: [-1.1075% -0.4105% +0.3172%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

U256::rotate_left       time:   [7.7551 ns 7.7892 ns 7.8216 ns]                               
                        change: [+0.7183% +1.2633% +1.7584%] (p = 0.00 < 0.05)
                        Change within noise threshold.

U256::rotate_right      time:   [6.8823 ns 6.8993 ns 6.9163 ns]                                
                        change: [-1.9033% -1.4820% -1.0720%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe